+++ src/constants.ts
@@ -1,0 +1,10 @@
+/**
+ * Radius of a single drone in meters in an outdoor setting.
+ */
+export const DEFAULT_DRONE_RADIUS = 0.75;
+
+/**
+ * Downscaling factor for indoor drones, based on the default drone radius
+ * for an outdoor setting.
+ */
+export const INDOOR_DRONE_SIZE_SCALING_FACTOR = 0.2;